requirements.txt
aiofiles==0.5.0
aiosqlite==0.15.0
alembic==1.6.5
anyio==3.3.0
asgiref==3.4.1
asyncpg==0.24.0
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
databases==0.3.2
ecdsa==0.17.0
fastapi==0.60.1
greenlet==1.1.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==3.2
importlib-metadata==4.6.3
iniconfig==1.1.1
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
packaging==21.0
passlib==1.7.4
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.4
pytest-asyncio==0.15.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.22
sqlalchemy2-stubs==0.0.2a19
sqlmodel==0.0.4
starlette==0.13.6
toml==0.10.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.6
uuid==1.30
uvicorn==0.14.0
websockets==8.1
zipp==3.5.0
